+++ packages/core/src/types.ts
@@ -126,6 +126,10 @@ export type ComponentRegistration = {
   };
 };
 
+export type ComponentRegistrationOptions = {
+  enabledBuiltInComponents?: string[];
+};
+
 export type Binding = {
   spaceId: string;
   environmentId: string;