+++ packages/experience-builder-sdk/src/blocks/preview/CompositionBlock.spec.tsx
@@ -2,10 +2,7 @@ import React from 'react';
 
 import { render } from '@testing-library/react';
 
-import {
-  CONTENTFUL_CONTAINER_ID,
-  CONTENTFUL_SECTION_ID,
-} from '@contentful/experience-builder-core/constants';
+import { CONTENTFUL_COMPONENTS } from '@contentful/experience-builder-core/constants';
 import { defineComponents, resetComponentRegistry } from '../../core/componentRegistry';
 import type { CompositionNode, ExperienceEntry } from '@contentful/experience-builder-core/types';
 import { CompositionBlock } from './CompositionBlock';
@@ -76,7 +73,7 @@ describe('CompositionBlock', () => {
 
   it('renders section node', () => {
     const sectionNode: CompositionNode = {
-      definitionId: CONTENTFUL_SECTION_ID,
+      definitionId: CONTENTFUL_COMPONENTS.section.id,
       variables: {},
       children: [],
     };
@@ -99,7 +96,7 @@ describe('CompositionBlock', () => {
 
   it('renders container node', () => {
     const containerNode: CompositionNode = {
-      definitionId: CONTENTFUL_CONTAINER_ID,
+      definitionId: CONTENTFUL_COMPONENTS.container.id,
       variables: {},
       children: [],
     };